Render video on its own framebuffer object
Before, the video from mpv was being rendered on the default
framebuffer, taking up the whole window. Now the video is drawn on a
separate framebuffer object and the texture is then drawn on the default
framebuffer. This allows the video to only cover a portion of the window
by changing the size of the framebuffer's texture, so there can be space
for the rest of the UI elements.